+++ dataland-community-manager/src/main/kotlin/org/dataland/datalandcommunitymanager/entities/MessageEntity.kt
@@ -90,7 +90,6 @@ data class MessageEntity(
                 )
                 companyOwnerList.map { TemplateEmailMessage.UserIdEmailRecipient(it.userId) }
             } else {
-
                 listOf()
             }
         }
